Move AbstractConnectableXml and its dependencies to public API and add AbstractConnectableXml::writeCurrentLimits which takes a prefix
