+++ include/powsybl/iidm/converter/xml/PropertiesXml.hpp
