+++ src/iidm/converter/xml/TieLineXml.hpp
@@ -11,8 +11,7 @@
 #include <powsybl/iidm/Network.hpp>
 #include <powsybl/iidm/TieLine.hpp>
 #include <powsybl/iidm/TieLineAdder.hpp>
-
-#include "AbstractConnectableXml.hpp"
+#include <powsybl/iidm/converter/xml/AbstractConnectableXml.hpp>
 
 namespace powsybl {
